pyproject.toml
[project]
name = "medical-summary-builder"
version = "0.1.0"
description = "LLM-powered pipeline that ingests medical case PDFs and produces structured medical summaries."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pdfplumber>=0.11",
    "python-docx>=1.1",
    "openai>=1.30",
    "python-dotenv>=1.0",
    "rich>=13.0",
    "click>=8.1",
    "pydantic>=2.7",
    "tenacity>=8.3",
    "rapidfuzz>=3.9",
    "pypdfium2>=5.7.0",
]

[project.scripts]
medical-summary = "medical_summary_builder.main:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/medical_summary_builder"]

[dependency-groups]
dev = [
    "pytest>=8.2",
    "pytest-cov>=5.0",
    "ruff>=0.4",
]
